ocaml: fix formatting

+++ pkgs/development/compilers/ocaml/generic.nix
@@ -93,9 +93,13 @@ stdenv.mkDerivation (
     patches = map fetchpatch' patches;
 
     # https://github.com/ocaml/ocaml/issues/14543
-    postPatch = if stdenv.cc.isClang then ''
-      rm testsuite/tests/basic/trigraph.ml
-    '' else null;
+    postPatch =
+      if stdenv.cc.isClang then
+        ''
+          rm testsuite/tests/basic/trigraph.ml
+        ''
+      else
+        null;
 
     strictDeps = true;
